#OSIS-665 Refactoring des forms pour généraliser
