+++ base/forms/learning_unit_search.py
@@ -1,0 +1,74 @@
+##############################################################################
+#
+#    OSIS stands for Open Student Information System. It's an application
+#    designed to manage the core business of higher education institutions,
+#    such as universities, faculties, institutes and professional schools.
+#    The core business involves the administration of students, teachers,
+#    courses, programs and so on.
+#
+#    Copyright (C) 2015-2018 Université catholique de Louvain (http://www.uclouvain.be)
+#
+#    This program is free software: you can redistribute it and/or modify
+#    it under the terms of the GNU General Public License as published by
+#    the Free Software Foundation, either version 3 of the License, or
+#    (at your option) any later version.
+#
+#    This program is distributed in the hope that it will be useful,
+#    but WITHOUT ANY WARRANTY; without even the implied warranty of
+#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#    GNU General Public License for more details.
+#
+#    A copy of this license - GNU General Public License - is available
+#    at the root of the source code of this program.  If not,
+#    see http://www.gnu.org/licenses/.
+#
+##############################################################################
+
+from django import forms
+from django.core.exceptions import ValidationError
+from django.db.models import Prefetch
+from django.utils.translation import ugettext_lazy as _
+
+from base import models as mdl
+from base.business.entity import get_entities_ids
+from base.business.entity_version import SERVICE_COURSE
+from base.business.learning_unit_year_with_context import append_latest_entities
+from base.forms.common import get_clean_data, treat_empty_or_str_none_as_none, TooManyResultsException
+from base.models import entity_version as mdl_entity_version, learning_unit_year
+from base.models.academic_year import AcademicYear, current_academic_year
+from base.models.enums import entity_container_year_link_type, learning_container_year_types, \
+    learning_unit_year_subtypes, active_status
+from base.forms.utils.uppercase import convert_to_uppercase
+
+
+class SearchForm(forms.Form):
+    MAX_RECORDS = 1000
+    ALL_LABEL = (None, _('all_label'))
+    ALL_CHOICES = (ALL_LABEL,)
+
+
+    academic_year_id = forms.ModelChoiceField(
+        label=_('academic_year_small'),
+        queryset=AcademicYear.objects.all(),
+        empty_label=_('all_label'),
+        required=False,
+    )
+
+    requirement_entity_acronym = forms.CharField(
+        max_length=20,
+        required=False,
+        label=_('requirement_entity_small')
+    )
+
+    acronym = forms.CharField(
+        max_length=15,
+        required=False,
+        label=_('acronym')
+    )
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.fields['academic_year_id'].initial = current_academic_year()
+
+    def clean_requirement_entity_acronym(self):
+        return convert_to_uppercase(self.cleaned_data.get('requirement_entity_acronym'))
