+++ base/forms/learning_units.py
@@ -35,57 +35,34 @@ from base.business.entity_version import SERVICE_COURSE
 from base.business.learning_unit_year_with_context import append_latest_entities
 from base.forms.common import get_clean_data, treat_empty_or_str_none_as_none, TooManyResultsException
 from base.models import entity_version as mdl_entity_version, learning_unit_year
-from base.models.academic_year import AcademicYear, current_academic_year
 from base.models.enums import entity_container_year_link_type, learning_container_year_types, \
     learning_unit_year_subtypes, active_status
-from base.forms.utils.uppercase import convert_to_uppercase
-
-MAX_RECORDS = 1000
-ALL_CHOICES = ((None, _('all_label')),)
+from base.forms.learning_unit_search import SearchForm
 
 
-class LearningUnitYearForm(forms.Form):
-    academic_year_id = forms.ModelChoiceField(
-        label=_('academic_year_small'),
-        queryset=AcademicYear.objects.all(),
-        empty_label=_('all_label'),
-        required=False
-    )
-
+class LearningUnitYearForm(SearchForm):
     container_type = forms.ChoiceField(
         label=_('type'),
-        choices=ALL_CHOICES + learning_container_year_types.LEARNING_CONTAINER_YEAR_TYPES,
+        choices=SearchForm.ALL_CHOICES + learning_container_year_types.LEARNING_CONTAINER_YEAR_TYPES,
         required=False
     )
 
     subtype = forms.ChoiceField(
         label=_('subtype'),
-        choices=ALL_CHOICES + learning_unit_year_subtypes.LEARNING_UNIT_YEAR_SUBTYPES,
+        choices=SearchForm.ALL_CHOICES + learning_unit_year_subtypes.LEARNING_UNIT_YEAR_SUBTYPES,
         required=False
     )
 
     status = forms.ChoiceField(
         label=_('status'),
-        choices=ALL_CHOICES + active_status.ACTIVE_STATUS_LIST[:-1],
+        choices=SearchForm.ALL_CHOICES + active_status.ACTIVE_STATUS_LIST[:-1],
         required=False
-    )
-
-    acronym = forms.CharField(
-        max_length=20,
-        required=False,
-        label=_('acronym')
     )
 
     title = forms.CharField(
         max_length=20,
         required=False,
         label=_('title')
-    )
-
-    requirement_entity_acronym = forms.CharField(
-        max_length=20,
-        required=False,
-        label=_('requirement_entity_small')
     )
 
     allocation_entity_acronym = forms.CharField(
@@ -99,7 +76,6 @@ class LearningUnitYearForm(forms.Form):
     def __init__(self, *args, **kwargs):
         self.service_course_search = kwargs.pop('service_course_search', False)
         super().__init__(*args, **kwargs)
-        self.fields['academic_year_id'].initial = current_academic_year()
 
     def clean_acronym(self):
         data_cleaned = self.cleaned_data.get('acronym')
@@ -107,9 +83,6 @@ class LearningUnitYearForm(forms.Form):
         if data_cleaned and learning_unit_year.check_if_acronym_regex_is_valid(data_cleaned) is None:
             raise ValidationError(_('LU_ERRORS_INVALID_REGEX_SYNTAX'))
         return data_cleaned
-
-    def clean_requirement_entity_acronym(self):
-        return convert_to_uppercase(self.cleaned_data.get('requirement_entity_acronym'))
 
     def clean_allocation_entity_acronym(self):
         data_cleaned = self.cleaned_data.get('allocation_entity_acronym')
@@ -119,7 +92,8 @@ class LearningUnitYearForm(forms.Form):
 
     def clean(self):
         if not self.service_course_search \
-                and self.cleaned_data and learning_unit_year.count_search_results(**self.cleaned_data) > MAX_RECORDS:
+                and self.cleaned_data \
+                and learning_unit_year.count_search_results(**self.cleaned_data) > SearchForm.MAX_RECORDS:
             raise TooManyResultsException
         return get_clean_data(self.cleaned_data)
